+++ Dental-House-UI/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="./public/images/logo.jpg" />
+    <link rel="icon" type="image/svg+xml" href="./public/images/logo.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Dental House</title>
     <link rel="preconnect" href="https://fonts.googleapis.com" />
